Refactors AppEnum to use a template mapper class
Replaces the internal EnumMapper with a template class
AppEnumMapper to allow for specialization per enum type.

This change improves type safety and avoids potential conflicts
when dealing with multiple enum types.
